design: 모바일 화면 시 제목에 패딩추가 , refactor: 파일명 TempData->PostData로 수정
